Allow pattern to be a function
In some cases, the pattern matching logic for which files to instrument
is more complicated than what can be easily added to a regular
expression. In these cases, you can set a function as the pattern. This
function will get executed with the filename and can perform any user
logic to determine if the file should be instrumented.