Exit instrumentation when user redefines "window"
The redefine of “window” in the middle of the function breaks code
coverage because it depends on “window”. The reason why it breaks it is
because JavaScript moves all variables to the top of the function. When
adding instrumentation, var'ing “window” will cause it to be null from
the start of the function that window._$blanket depends on.

Fixes bug #254