Issue #258: Added phantomjs as an npm dependency so that new developers do not have to manually install phantom to get npm test to work.
Changes Include:

* package.json

Added a dev dependency to phantoms v1.8.2-0. This version was explicitly chosen because of a bug that occurs during the install process for phantomjs where the unzipped binary does not get moved to the right location. For more information on this bug, you can see it here: https://github.com/Obvious/phantomjs/issues/38

* test/testconfigs.json

Changed the "phantom" command to not rely on phantomjs to be on the PATH, it will instead reference the phantomjs that gets installed into node_modules by npm.