Add per-module totals in QUnit reporter
In addition to the grand total already available on the last line of
the report, this change allows a user to specify a regular expression
which describes how modules are named in your application file
structure.  When provided, the QUnit reporter also outputs the
per-module totals in addition to the grand total.