Internet Explorer console color support
As setting the colors on IE's console is not an option, we can use
console.info for good messages, and console.error for errors.

As Internet Explorer only returns info about its console when the dev
tools are active, it can be helpful to have the console open before
loading the page. Either that, or you can reload the page after opening
the console.